refactor: extract job catalog gate helpers
